+++ src/Services/Strategies/Contracts/DiscountStrategyContract.php
@@ -4,5 +4,5 @@ namespace EscolaLms\Cart\Services\Strategies\Contracts;
 
 interface DiscountStrategyContract
 {
-    public function total(float $subtotal, float $tax): float;
+    public function total(int $subtotal, int $tax): int;
 }
