+++ src/Services/Strategies/Discount/AmountStrategy.php
@@ -6,7 +6,7 @@ use EscolaLms\Cart\Services\Strategies\Abstracts\DiscountStrategy;
 
 class AmountStrategy extends DiscountStrategy
 {
-    public function total(float $subtotal, float $tax): float
+    public function total(int $subtotal, int $tax): int
     {
         return $this->value(($subtotal + $tax) - $this->discount->value);
     }
